Fixed DCAS not being lock-free on 32-bit x86 target with clang.
Clang has a bug of not advertising support for atomics implemented via
cmpxchg8b, even if the instruction is enabled in the command line. We have
to workaround the same problem with cmpxchg16b on 64-bit x86 as well, so
we apply the same approach here - we implement all atomic ops through DCAS
ourselves.

This fix adds a check for cmpxchg8b to capabilities definition.