Added a workaround for MSVC 14.0 alignas issues in 32-bit mode.
The compiler allows to apply alignas but later fails to pass arguments
of the aligned types to functions with error C2719. At the same time,
std::max_align_t has alignment of 8 and the error doesn't show up when
the type is aligned using the union trick. Thus we disable alignas
for MSVC 14.0 in 32-bit mode.

Also, use std::max_align_t on MSVC, when possible.