Documented value() operation, clarified the limitation of no padding bits.
The value() operation is useful with futexes, but should not be used for
anything else, basically.

The lack of support for types with padding bits is documented more prominently.
The docs do mention that `long double` on x86 is supported though.

Also, added description of the new tests added recently.

Related to https://github.com/boostorg/atomic/issues/34.