Use api-ms-win-core-synch-l1-2-0.dll to query WaitOnAddress and friends.
It was suggested in a comment[1] that the correct dll to use to resolve
WaitOnAddress, WakeByAddressSingle and WakeByAddressAll is
api-ms-win-core-synch-l1-2-0.dll instead of KernelBase.dll. On some systems
KernelBase.dll may not be available and the WaitOnAddress API may be implemented
in a different library.

Tests have shown that GetModuleHandleW(L"api-ms-win-core-synch-l1-2-0.dll")
returns a handle for KernelBase.dll anyway. Also, there exists a presumably
older version of this library: api-ms-win-core-synch-l1-1-0.dll. The older
version is also "loaded" into the process and also resolves to KernelBase.dll,
which suggests that hopefully api-ms-win-core-synch-l1-2-0.dll will stay
available and working in the forseeable future.

[1]: https://github.com/microsoft/STL/pull/593#issuecomment-641019129