Drop checks for BOOST_NO_ALIGNMENT and BOOST_HAS_INT128 for x86.
We no longer use the alignment attributes (except for alignas, when
available) in order to align the 128-bit storage for atomics. Instead we
rely on type_with_alignment for that. Although it may still use the same
attributes to acheve the required alignment, this is its implementation
detail and may not correspond to BOOST_NO_ALIGNMENT exactly.

The check for BOOST_HAS_INT128 was not really relevant to begin with
because __int128 is not guaranteed to have alignment of 16.

In any case, all current compilers targeting x86 do support alignment of
16, so the checks weren't doing anything.