Suppress default constructor of the target type in bitwise_cast.
Since the default constructor may be non-trivial and throwing, we have
to suppress it in the memcpy-based implementation of bitwise_cast. Since
the type must be trivially copyable, initializing the object using memcpy
is a valid way of creating it, so there is no UB.

This is necessary for all atomic operations that involve bitwise_cast
internally to remain noexcept, as specified in the C++ standard.

Also removed remove_cv shim since we now require type_traits and the
shim is no longer used.