Removed the check for _M_ARMT macro.
The macro is not documented in MSDN and it's not entirely clear when
it is defined (and, in particular, whether _M_ARM is also defined at
the same time). Remove it for now, until there appears demand for
it and a knowledgeable user who knows what it means and when it is
defined.