Use a more unique string for type_traits config check.
This avoids the potential clash with another library that could make
a similar check for <type_traits> but require a different set of
type traits. b2 seems to use the string as part of the key in the
config cache, so if the two different config checks have a matching
description string it is unknown which check result is cached and used.